Refactor evaluation from the Benchmark to the BenchmarkTask object.
This ensures that task evaluation is more foolproof. For example, this ensures that a user will not accidentally evaluate their task on the wrong split of data. It also allows users to use several heterogeneous Tasks within the same Benchmark subclass.

IMPORTANT: This change also updates NitroML-curated BenchmarkTasks to use the 'test' split for evaluation, when available. Therefore benchmark results following this CL for several BenchmarkTasks will be incomparable with results from previous runs.
PiperOrigin-RevId: 339774218