Allow users to individually customize the pipeline_run_id used to resolve each input channel.
When pipeline_filtering is used on an existing pipeline, a component could have its upstream components filtered out. This causes the component to have an input channel that can't produce an artifact. The way we have been handling this is to allow the user to specify an old_pipeline_run_id that the input channel would use to query for artifacts.

However, sometimes when a user runs a sequence of partial runs, they may find that they want to have different input channels querying from different past pipeline runs. This change enables the user to supply a `pipeline_run_id_fn`, so they can customize how to map each input channel to the desired pipeline_run_id.

PiperOrigin-RevId: 357023684