Support for raw prediction evaluation in BenchmarkTask.
Previously, in `BenchmarkTask#make_evaluation`, NitroML BenchmarkTask ask for a SavedModel channel and does the inference and metric calculation as a part of the "evaluation" step.

There are a few use cases where accepting a SavedModel from users is not appropriate:
- When inference computation is a custom process, needing to be handled and set up by user (eg. bringing up an embedding-lookup server).
- When a user is not using TF and so their trainer does not emit a SavedModel.

To support those use-cases, `BenchmarkTask#make_evaluation` now accepts raw predictions as a channel of Examples containing the raw predictions.

To ensure fair comparisons between evaluations made on SavedModel predictions and raw predictions, TFMA is now configured to treat the SavedModel as a `tf_generic` model, meaning that metrics computed by the model will be ignored.

PiperOrigin-RevId: 358025414